# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POWERLVL v1 API contract — anonymous and member scans, public feed, leaderboards, profiles, and account.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    post:
      operationId: createScan
      tags: [scans]
      summary: Create a Scan
      description: |
        Multipart upload that runs the full Scanner pipeline server-side:
        validation, moderation, image processing, Vision Model analysis,
        slop detection, scoring, anomaly draw, and persistence. Available to
        both Anonymous Users (within their daily quota) and Members.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
            encoding:
              image:
                contentType: image/jpeg, image/png, image/heic, image/webp
              category:
                contentType: text/plain
              description:
                contentType: text/plain
              localDate:
                contentType: text/plain
              localTz:
                contentType: text/plain
      responses:
        "200":
          description: Scan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateScanResponse"
        "400":
          description: Invalid input, image rejected, or moderation flagged
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ApiErrorInvalidInput"
                  - $ref: "#/components/schemas/ApiErrorImageInvalid"
                  - $ref: "#/components/schemas/ApiErrorModerationRejected"
                discriminator:
                  propertyName: code
                  mapping:
                    INVALID_INPUT: "#/components/schemas/ApiErrorInvalidInput"
                    IMAGE_INVALID: "#/components/schemas/ApiErrorImageInvalid"
                    MODERATION_REJECTED: "#/components/schemas/ApiErrorModerationRejected"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          description: Rate limit or daily limit reached
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ApiErrorRateLimited"
                  - $ref: "#/components/schemas/ApiErrorDailyLimitReached"
                discriminator:
                  propertyName: code
                  mapping:
                    RATE_LIMITED: "#/components/schemas/ApiErrorRateLimited"
                    DAILY_LIMIT_REACHED: "#/components/schemas/ApiErrorDailyLimitReached"
        "503":
          description: Vision or moderation service unavailable
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ApiErrorModerationUnavailable"
                  - $ref: "#/components/schemas/ApiErrorVisionUnavailable"
                discriminator:
                  propertyName: code
                  mapping:
                    MODERATION_UNAVAILABLE: "#/components/schemas/ApiErrorModerationUnavailable"
                    VISION_UNAVAILABLE: "#/components/schemas/ApiErrorVisionUnavailable"
        "500":
          $ref: "#/components/responses/Internal"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get a Scan
      description: |
        Returns the public projection of a Scan. Used by the permalink page
        and Feed entry detail. Includes signed image URLs valid for ≤ 24 h
        and the absolute permalink plus the three share-card URLs.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanPublic"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /scans/{id}/share-card.png:
    get:
      operationId: getScanShareCard
      tags: [scans]
      summary: Get a Scan's share-card PNG
      description: |
        Renders the share card for the requested Scan at the requested aspect
        ratio. Delegates to `artifacts/share-card`. Cached at the CDN with
        `public, max-age=86400, s-maxage=86400, immutable`.
      parameters:
        - $ref: "#/components/parameters/ScanId"
        - name: ratio
          in: query
          required: true
          description: Share-card aspect ratio
          schema:
            $ref: "#/components/schemas/ShareCardRatio"
      responses:
        "200":
          description: Rendered share card
          headers:
            Cache-Control:
              schema:
                type: string
              description: "public, max-age=86400, s-maxage=86400, immutable"
            ETag:
              schema:
                type: string
              description: sha256(scanId + ratio + version)
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          description: Share-card renderer temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorInternal"
  /scans/{id}/claim:
    post:
      operationId: claimScan
      tags: [scans]
      summary: Claim an unclaimed Scan
      description: |
        Soft-auth claim. Requires an authenticated Member. Verifies the Scan
        has `member_id IS NULL` and that its `anon_session_id` matches the
        caller's `plvl_anon` cookie. Idempotent for a Scan already owned by
        the caller; returns 403 when owned by another Member.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan claimed (or already owned by caller)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimScanResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  parameters:
    ScanId:
      name: id
      in: path
      required: true
      description: Scan UUID
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthenticated:
      description: Caller is not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorUnauthenticated"
    Forbidden:
      description: Caller is not permitted to access the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorForbidden"
    NotFound:
      description: Resource does not exist or has been expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorNotFound"
    RateLimited:
      description: Caller exceeded a rate limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorRateLimited"
    Internal:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorInternal"
  schemas:
...
    Category:
      type: string
      description: Scan Category. Determines the five Category Stats.
      enum:
        - SETUPS
        - FITNESS
        - DRIP
        - PETS
        - RIDES
        - WILDCARD
    Tier:
      type: string
      description: Tier bucket derived from the final Score.
      enum:
        - D
        - C
        - B
        - A
        - S
        - SS
        - SSS
        - LIMITLESS
    AnomalyType:
      type: string
      description: Rare reveal event assigned to some Scans. At most one per Scan.
      enum:
        - POWER_SURGE_DETECTED
        - FORBIDDEN_AURA
        - SCOUTER_FAILURE
        - UNREGISTERED_ENERGY
        - CHAOS_SPIKE
    RevealVariant:
      type: string
      description: Client reveal variant the web client must play. 1:1 with `AnomalyType` (plus `standard` when no anomaly).
      enum:
        - standard
        - power_surge
        - forbidden_aura
        - scouter_failure
        - unregistered_energy
        - chaos_spike
    ShareCardRatio:
      type: string
      description: Share-card aspect ratio.
      enum:
        - story
        - square
        - landscape
...
    CoreStats:
      type: object
      description: The four Core Stats present on every Scan, each in the inclusive range 0..10000.
      additionalProperties: false
      properties:
        aura:
          type: integer
          minimum: 0
          maximum: 10000
        power:
          type: integer
          minimum: 0
          maximum: 10000
        status:
          type: integer
          minimum: 0
          maximum: 10000
        threat:
          type: integer
          minimum: 0
          maximum: 10000
      required: [aura, power, status, threat]
    CategoryStats:
      type: object
      description: |
        The five Category-specific stats produced for a Scan, each in the inclusive
        range 0..10000. Keys depend on the Scan's Category and are stable per
        Category (see design.md "Per-Category Stats"). Object has exactly five keys.
      additionalProperties:
        type: integer
        minimum: 0
        maximum: 10000
    ShareCardUrls:
      type: object
      description: Absolute URLs for the three share-card aspect ratios.
      additionalProperties: false
      properties:
        story:
          type: string
          format: uri
        square:
          type: string
          format: uri
        landscape:
          type: string
          format: uri
      required: [story, square, landscape]
    ScanPublic:
      type: object
      description: Public projection of a Scan, returned by Feed, Permalink, and Profile endpoints.
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        username:
          type: [string, "null"]
          description: Member username, or null when the Scan is unclaimed (anonymous).
          pattern: "^[a-z0-9_]{3,20}$"
        category:
          $ref: "#/components/schemas/Category"
        score:
          type: integer
          minimum: 1000
          maximum: 100000
        tier:
          $ref: "#/components/schemas/Tier"
        verdictNoun:
          type: string
          description: Verdict noun selected from the tier pool.
          minLength: 1
        coreStats:
          $ref: "#/components/schemas/CoreStats"
        categoryStats:
          $ref: "#/components/schemas/CategoryStats"
        commentary:
          type: string
          description: Final commentary line, 8..14 words inclusive.
          minLength: 1
        description:
          type: [string, "null"]
          description: Sanitized user-supplied description, ≤ 120 chars; null when absent.
          maxLength: 120
        anomaly:
          oneOf:
            - $ref: "#/components/schemas/AnomalyType"
            - type: "null"
        imageUrl:
          type: string
          format: uri
          description: Signed URL for the processed image (≤ 24 h validity).
        thumbnailUrl:
          type: string
          format: uri
          description: Signed URL for the 480px thumbnail (≤ 24 h validity).
        createdAt:
          type: string
          format: date-time
        permalink:
          type: string
          format: uri
          description: Absolute URL of the Scan permalink (`/scan/{id}`).
        shareCardUrls:
          $ref: "#/components/schemas/ShareCardUrls"
        likeCount:
          type: integer
          minimum: 0
        likedByMe:
          type: boolean
      required:
        - id
        - username
        - category
        - score
        - tier
        - verdictNoun
        - coreStats
        - categoryStats
        - commentary
        - description
        - anomaly
        - imageUrl
        - thumbnailUrl
        - createdAt
        - permalink
        - shareCardUrls
        - likeCount
        - likedByMe
...
    CreateScanRequest:
      type: object
      description: Multipart/form-data payload for creating a Scan.
      additionalProperties: false
      properties:
        image:
          type: string
          format: binary
          description: Image bytes. JPEG, PNG, HEIC, or WebP. ≤ 8 MB. Shorter edge ≥ 256 px.
        category:
          $ref: "#/components/schemas/Category"
        description:
          type: string
          description: Optional user-supplied description (≤ 120 chars). Sanitized server-side.
          maxLength: 120
        localDate:
          type: string
          description: Visitor's local date in `YYYY-MM-DD` format. Used for the anonymous daily-limit window.
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        localTz:
          type: string
          description: IANA timezone string (optional, refines the anonymous daily window).
      required: [image, category, localDate]
    CreateScanRevealHints:
      type: object
      description: Reveal-sync hints returned alongside a created Scan.
      additionalProperties: false
      properties:
        serverElapsedMs:
          type: integer
          minimum: 0
          description: Milliseconds elapsed in the pipeline before the response was emitted.
        revealVariant:
          $ref: "#/components/schemas/RevealVariant"
      required: [serverElapsedMs, revealVariant]
    CreateScanResponse:
      type: object
      additionalProperties: false
      properties:
        scan:
          $ref: "#/components/schemas/ScanPublic"
        reveal:
          $ref: "#/components/schemas/CreateScanRevealHints"
      required: [scan, reveal]
    ClaimScanResponse:
      type: object
      additionalProperties: false
      properties:
        scan:
          $ref: "#/components/schemas/ScanPublic"
      required: [scan]
...
    ApiErrorInvalidInput:
      type: object
      description: Request payload failed validation.
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [INVALID_INPUT]
        message:
          type: string
        field:
          type: string
          description: Optional name of the offending field.
        rule:
          type: string
          description: Optional name of the violated rule (e.g. `length`, `chars`, `uniqueness`).
      required: [code, message]
    ApiErrorUnauthenticated:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [UNAUTHENTICATED]
        message:
          type: string
      required: [code, message]
    ApiErrorForbidden:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [FORBIDDEN]
        message:
          type: string
      required: [code, message]
    ApiErrorNotFound:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [NOT_FOUND]
        message:
          type: string
      required: [code, message]
    ApiErrorRateLimited:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [RATE_LIMITED]
        message:
          type: string
        retryAfterSec:
          type: integer
          minimum: 0
      required: [code, message, retryAfterSec]
    ApiErrorDailyLimitReached:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [DAILY_LIMIT_REACHED]
        message:
          type: string
        resetAt:
          type: string
          format: date-time
          description: Next reset time (visitor's next local-day for anonymous, next UTC midnight for Member).
      required: [code, message, resetAt]
    ApiErrorModerationRejected:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [MODERATION_REJECTED]
        message:
          type: string
        categories:
          type: array
          items:
            type: string
          description: Moderation categories triggered (e.g. `Explicit Nudity`, `Violence`, `Hate Symbols`).
      required: [code, message, categories]
    ApiErrorModerationUnavailable:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [MODERATION_UNAVAILABLE]
        message:
          type: string
      required: [code, message]
    ApiErrorImageInvalid:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [IMAGE_INVALID]
        message:
          type: string
        reason:
          type: string
          enum: [size, format, dimensions, mime_mismatch]
      required: [code, message, reason]
    ApiErrorVisionUnavailable:
      type: object
      description: |
        Surfaces when the Vision Model is unrecoverable for this request.
        On the standard Scanner pipeline the failure path falls through to a
        SCOUTER_FAILURE neutral fallback rather than returning this error;
        this code exists for future explicit handling and documentation.
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [VISION_UNAVAILABLE]
        message:
          type: string
      required: [code, message]
    ApiErrorInternal:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [INTERNAL]
        message:
          type: string
      required: [code, message]
...